Moved JDK 17 Handler to core, removed JDK 17 package
