--- persistence/binary-jdk17/pom.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-	<modelVersion>4.0.0</modelVersion>
-
-	<parent>
-		<groupId>org.eclipse.serializer</groupId>
-		<artifactId>persistence-parent</artifactId>
-		<version>3.0.0-SNAPSHOT</version>
-		<relativePath>../pom.xml</relativePath>
-	</parent>
-
-	<artifactId>persistence-binary-jdk17</artifactId>
-	
-	<name>Eclipse Serializer Persistence JDK17</name>
-	<description>Eclipse Serializer Persistence Specialized Type Handlers for JDK17+</description>
-	<url>https://projects.eclipse.org/projects/technology.serializer</url>
-	
-	<properties>
-		<maven.compiler.source>17</maven.compiler.source>
-		<maven.compiler.target>17</maven.compiler.target>
-	</properties>
-	
-	<dependencies>
-		<dependency>
-			<groupId>org.eclipse.serializer</groupId>
-			<artifactId>persistence-binary</artifactId>
-			<version>${project.version}</version>
-		</dependency>
-	</dependencies>
-
-</project>
