--- persistence/binary-jdk17/src/main/java/module-info.java
@@ -1,23 +1,0 @@
-/*-
- * #%L
- * Eclipse Serializer Persistence JDK17
- * %%
- * Copyright (C) 2023 MicroStream Software
- * %%
- * This program and the accompanying materials are made
- * available under the terms of the Eclipse Public License 2.0
- * which is available at https://www.eclipse.org/legal/epl-2.0/
- * 
- * SPDX-License-Identifier: EPL-2.0
- * #L%
- */
-module org.eclipse.serializer.persistence.binary.jdk17
-{
-	exports org.eclipse.serializer.persistence.binary.jdk17.java.util;
-	exports org.eclipse.serializer.persistence.binary.jdk17.types;
-
-	requires transitive org.eclipse.serializer.persistence.binary;
-	/* To indicate we use these modules directly, but they also come through org.eclipse.serializer.persistence.binary */
-	requires org.eclipse.serializer.base;
-	requires org.eclipse.serializer.persistence;
-}
