+++ persistence/binary/src/main/java/org/eclipse/serializer/persistence/binary/java/util/BinaryHandlerBitSet.java
@@ -1,5 +1,19 @@
 package org.eclipse.serializer.persistence.binary.java.util;
 
+/*-
+ * #%L
+ * Eclipse Serializer Persistence Binary
+ * %%
+ * Copyright (C) 2023 - 2025 MicroStream Software
+ * %%
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ * 
+ * SPDX-License-Identifier: EPL-2.0
+ * #L%
+ */
+
 import java.util.BitSet;
 
 import org.eclipse.serializer.memory.XMemory;
