+++ persistence/pom.xml
@@ -27,15 +27,4 @@
         <module>binary-android</module>
     </modules>
 
-    <profiles>
-        <profile>
-            <id>from_java_17</id>
-            <activation>
-                <jdk>[17,)</jdk>
-            </activation>
-            <modules>
-                <module>binary-jdk17</module>
-            </modules>
-        </profile>
-    </profiles>
 </project>
